Admin 고객관리 완료, 계좌관리 진행중

+++ src/main/java/com/spring/domain/DepositVO.java
@@ -10,11 +10,11 @@ public class DepositVO {
 	private int type;
 	private int product;
 	private String password;
-	private int balance;
+	private long balance;
 	private int card_apply;
 	private String card_no;
-	private int day_withdraw;
-	private int max_withdraw;
+	private long day_withdraw;
+	private long max_withdraw;
 	private String branch;
 	private Date createdate;
 	private Date updatedate;
